Storing all raw data as integers
This is not the correct way. Double and Long values will lose data and char and string storage is inefficient. But Scala's static type inference is dumb and does not let me do what I want to do.
When I know the type of an object at runtime, I want to cast other objects that I am sure are of the same type to it.