Adding a class to convert MathExpr to C/OpenCL.
The agg_cal call gets an array of struct as input. Each element of the
array represents a mathematical expression in an aggregation. We need to
constructs this array in Java. The java objects cannot be serialized
directly to cunstruct these types. So I am using a custom byte buffer to
enfore LITTLE_ENDIAN encoding of int and long values into the
expressions.

#30: port aggregations from opencl/cuda implementation

Task-Url: http://github.com/meisam/spark/issues/issue/30